Rhasspy Docker Compose (#39)
Add the necessary Docker compose file for creating Rhasspy instead of creating the container with a docker run command. Utilizing the docker compose file is important as it defines a Docker network utilized by Rhasspy and Sandman Web so that Sandman Web can better communicate with the Rhasspy container. Update the readme to reflect usage of Docker compose.